chore(release): update library versions and add repo version file
Updated the firmware versions in magis-lib and magisv2-lib to 2.2.2-beta and 2.4.2-beta, respectively, indicating minor changes or bug fixes. Additionally, added a new version file to track the overall repository version, starting with 2.2.0. This helps in maintaining consistency and clarity on version tracking across different components of the project.